% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qccheck.R
\name{check_logic}
\alias{check_logic}
\title{Logic checker}
\usage{
check_logic(Data, CorrelationThreshold = 0.5, ZThreshold = 2.3)
}
\arguments{
\item{Data}{A dataframe, ideally no more than 3000 columns (otherwise)
the algorithm will take a really long time to run}

\item{CorrelationThreshold}{A numeric, the correlation threshold}

\item{ZThreshold}{A numeric, the location in the similarity distribution
to flag respondents as being suspiciously similar (a z-score)}
}
\description{
\code{check_similarity} flags respondents whose response patterns
deviate too far from expected (highly correlated) patterns.
}

