% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qccheck.R
\name{check_similarity}
\alias{check_similarity}
\title{Similarity checker}
\usage{
check_similarity(Data, Condition1VarInData = NULL,
  Condition2VarInData = NULL, ZThreshold)
}
\arguments{
\item{Data}{A dataframe, ideally no more than 3000 columns (otherwise)
the algorithm will take a really long time to run}

\item{Condition1VarInData}{A character, identifying the first dimension
to the cut the similarity test on (optional)}

\item{Condition2VarInData}{A character, identifying the second dimension
to the cut the similarity test on (optional)}

\item{ZThreshold}{A numeric, the location in the similarity distribution
to flag respondents as being suspiciously similar (a z-score)}
}
\description{
\code{check_similarity} flags respondents whose response patterns
are suspiciously similar.
}

